requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-ajax-selects==1.7.1
django-attachments==1.3
django-ckeditor==5.6.1
django-guardian==1.4.9
django-heroku==0.3.1
django-js-asset==1.1.0
django-jsonfield==1.0.1
django-mptt==0.9.1
django-notify-x==0.1.9
django-reversion==3.0.2
django-session-security==2.6.5
django-simple-history==2.6.0
fake-factory==9999.9.9
Faker==1.0.1
gunicorn==19.9.0
lazy-object-proxy==1.3.1
Pillow==5.4.1
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.9
six==1.11.0
text-unidecode==1.2
typed-ast==1.1.0
whitenoise==4.1.2
wrapt==1.10.11
